+++ src/main/java/first/resale/models/Seller.java
@@ -1,8 +1,8 @@
 package first.resale.models;
 
 public class Seller {
-    private int id;
-    private int userId;
+    private long id;
+    private long userId;
     private String brandName;
     private String additionalPhone;
 }
